move secret recovery key into sdk
