+++ src/sdk/key.rs
@@ -1,0 +1,17 @@
+use std::{
+    array::TryFromSliceError,
+    convert::{TryFrom, TryInto},
+};
+use zeroize::Zeroize;
+
+#[derive(Debug, Clone, Zeroize)]
+#[zeroize(drop)]
+pub struct SecretRecoveryKey(pub(crate) [u8; 64]);
+
+impl TryFrom<&[u8]> for SecretRecoveryKey {
+    type Error = TryFromSliceError;
+
+    fn try_from(v: &[u8]) -> Result<Self, Self::Error> {
+        Ok(Self(v.try_into()?))
+    }
+}
