+++ src/sdk/mod.rs
@@ -1,3 +1,5 @@
 pub mod api;
 
+pub mod key;
+
 pub(crate) mod wire_bytes;
